requirements.txt
dragonfly2>=0.29.0
wxpython==4.0.7
pillow>=6.2.2
tomlkit>=0.5.11
future>=0.18.2
mock>=3.0.5
appdirs>=1.4.3
scandir>=1.10.0
pyvda==0.0.8
six